fix race condition & refactor details._initialize(done)
